wrong type!

+++ prot.h
@@ -36,7 +36,7 @@ struct RpHdr {
 struct OpOpen {
     char filename[FILENAME_LEN];
     uint8_t flags;
-    uint32_t mode;
+    uint16_t mode;
 };
 
 struct RpOpen {
